pyproject.toml
[project]
name = "dotalk"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=5.1.6",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "python-dotenv>=1.0.0",
    "drf-spectacular>=0.29.0",
    "google-auth>=2.48.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "redis>=7.2.1",
    "requests>=2.32.5",
]
